Refactor to extract test code for reuse
