+++ tests/api2/test_261_iscsi_cmd.py
@@ -12,6 +12,10 @@ import iscsi
 import pyscsi
 import pytest
 import requests
+from assets.websocket.iscsi import (alua_enabled, initiator, initiator_portal,
+                                    portal, read_capacity16, target,
+                                    target_extent_associate, verify_capacity,
+                                    verify_luns)
 from middlewared.service_exception import ValidationError, ValidationErrors
 from middlewared.test.integration.assets.iscsi import target_login_test
 from middlewared.test.integration.assets.pool import dataset, snapshot
@@ -20,11 +24,10 @@ from middlewared.test.integration.utils.client import truenas_server
 from pyscsi.pyscsi.scsi_sense import sense_ascq_dict
 from pytest_dependency import depends
 
-from auto_config import ha, hostname, isns_ip, pool_name, password, user
+from auto_config import ha, hostname, isns_ip, password, pool_name, user
 from functions import SSH_TEST
 from protocols import (initiator_name_supported, iscsi_scsi_connection,
                        isns_connection)
-
 
 # Setup some flags that will enable/disable tests based upon the capabilities of the
 # python-scsi package in use
@@ -134,51 +137,6 @@ def iscsi_auth(tag, user, secret, peeruser=None, peersecret=None):
 
 
 @contextlib.contextmanager
-def initiator(comment='Default initiator', initiators=[]):
-    payload = {
-        'comment': comment,
-        'initiators': initiators,
-    }
-    initiator_config = call('iscsi.initiator.create', payload)
-
-    try:
-        yield initiator_config
-    finally:
-        call('iscsi.initiator.delete', initiator_config['id'])
-
-
-@contextlib.contextmanager
-def portal(listen=[{'ip': '0.0.0.0'}], comment='Default portal', discovery_authmethod='NONE'):
-    payload = {
-        'listen': listen,
-        'comment': comment,
-        'discovery_authmethod': discovery_authmethod
-    }
-    portal_config = call('iscsi.portal.create', payload)
-
-    try:
-        yield portal_config
-    finally:
-        call('iscsi.portal.delete', portal_config['id'])
-
-
-@contextlib.contextmanager
-def target(target_name, groups, alias=None):
-    payload = {
-        'name': target_name,
-        'groups': groups,
-    }
-    if alias:
-        payload.update({'alias': alias})
-    target_config = call('iscsi.target.create', payload)
-
-    try:
-        yield target_config
-    finally:
-        call('iscsi.target.delete', target_config['id'], True)
-
-
-@contextlib.contextmanager
 def file_extent(pool_name, dataset_name, file_name, filesize=MB_512, extent_name='extent', serial=None):
     payload = {
         'type': 'FILE',
@@ -266,37 +224,6 @@ def zvol_extent(zvol, extent_name='zvol_extent'):
         yield extent_config
     finally:
         call('iscsi.extent.delete', extent_config['id'], True, True)
-
-
-@contextlib.contextmanager
-def target_extent_associate(target_id, extent_id, lun_id=0):
-    alua_enabled = call('iscsi.global.alua_enabled')
-    payload = {
-        'target': target_id,
-        'lunid': lun_id,
-        'extent': extent_id
-    }
-    associate_config = call('iscsi.targetextent.create', payload)
-    if alua_enabled:
-        # Give a little time for the STANDBY target to surface
-        sleep(2)
-
-    try:
-        yield associate_config
-    finally:
-        call('iscsi.targetextent.delete', associate_config['id'], True)
-    if alua_enabled:
-        sleep(2)
-
-
-@contextlib.contextmanager
-def initiator_portal():
-    with initiator() as initiator_config:
-        with portal() as portal_config:
-            yield {
-                'initiator': initiator_config,
-                'portal': portal_config,
-            }
 
 
 @contextlib.contextmanager
@@ -404,22 +331,6 @@ def isns_enabled(delay=5):
             sleep(delay)
 
 
-@contextlib.contextmanager
-def alua_enabled(delay=10):
-    payload = {'alua': True}
-    call('iscsi.global.update', payload)
-    if delay:
-        sleep(delay)
-        call('iscsi.alua.wait_for_alua_settled', 5, 8)
-    try:
-        yield
-    finally:
-        payload = {'alua': False}
-        call('iscsi.global.update', payload)
-        if delay:
-            sleep(delay)
-
-
 def TUR(s):
     """
     Perform a TEST UNIT READY.
@@ -474,71 +385,6 @@ def _verify_inquiry(s):
     data = r.result
     assert data['t10_vendor_identification'].decode('utf-8').startswith("TrueNAS"), str(data)
     assert data['product_identification'].decode('utf-8').startswith("iSCSI Disk"), str(data)
-
-
-def _extract_luns(rl):
-    """
-    Return a list of LUNs.
-
-    :param rl: a ReportLuns instance (response)
-    :return result a list of int LUNIDs
-
-    Currently the results from pyscsi.ReportLuns.unmarshall_datain are (a) subject
-    to change & (b) somewhat lacking for our purposes.  Therefore we will parse
-    the datain here in a manner more useful for us.
-    """
-    result = []
-    # First 4 bytes are LUN LIST LENGTH
-    lun_list_length = int.from_bytes(rl.datain[:4], "big")
-    # Next 4 Bytes are RESERVED
-    # Remaining bytes are LUNS (8 bytes each)
-    luns = rl.datain[8:]
-    assert len(luns) >= lun_list_length
-    for i in range(0, lun_list_length, 8):
-        lun = luns[i: i + 8]
-        addr_method = (lun[0] >> 6) & 0x3
-        assert addr_method == 0, f"Unsupported Address Method: {addr_method}"
-        if addr_method == 0:
-            # peripheral device addressing method, don't care about bus.
-            result.append(lun[1])
-    return result
-
-
-def _verify_luns(s, expected_luns):
-    """
-    Verify that the supplied SCSI has the expected LUNs.
-
-    :param s: a pyscsi.SCSI instance
-    :param expected_luns: a list of int LUNIDs
-    """
-    TUR(s)
-    # REPORT LUNS
-    rl = s.reportluns()
-    data = rl.result
-    assert isinstance(data, dict), data
-    assert 'luns' in data, data
-    # Check that we only have LUN 0
-    luns = _extract_luns(rl)
-    assert len(luns) == len(expected_luns), luns
-    assert set(luns) == set(expected_luns), luns
-
-
-def _read_capacity16(s):
-    # READ CAPACITY (16)
-    data = s.readcapacity16().result
-    return (data['returned_lba'] + 1 - data['lowest_aligned_lba']) * data['block_length']
-
-
-def _verify_capacity(s, expected_capacity):
-    """
-    Verify that the supplied SCSI has the expected capacity.
-
-    :param s: a pyscsi.SCSI instance
-    :param expected_capacity: an int
-    """
-    TUR(s)
-    returned_size = _read_capacity16(s)
-    assert returned_size == expected_capacity
 
 
 def get_target(targetid):
@@ -630,14 +476,14 @@ def test_02_read_capacity16(request):
                         with target_extent_associate(target_id, extent_id):
                             iqn = f'{basename}:{target_name}'
                             with iscsi_scsi_connection(truenas_server.ip, iqn) as s:
-                                _verify_capacity(s, MB_100)
+                                verify_capacity(s, MB_100)
                     # 512 MB file extent
                     with file_extent(pool_name, dataset_name, file_name, MB_512) as extent_config:
                         extent_id = extent_config['id']
                         with target_extent_associate(target_id, extent_id):
                             iqn = f'{basename}:{target_name}'
                             with iscsi_scsi_connection(truenas_server.ip, iqn) as s:
-                                _verify_capacity(s, MB_512)
+                                verify_capacity(s, MB_512)
                 # 100 MB zvol extent
                 with zvol_dataset(zvol, MB_100):
                     with zvol_extent(zvol) as extent_config:
@@ -645,7 +491,7 @@ def test_02_read_capacity16(request):
                         with target_extent_associate(target_id, extent_id):
                             iqn = f'{basename}:{target_name}'
                             with iscsi_scsi_connection(truenas_server.ip, iqn) as s:
-                                _verify_capacity(s, MB_100)
+                                verify_capacity(s, MB_100)
                 # 512 MB zvol extent
                 with zvol_dataset(zvol):
                     with zvol_extent(zvol) as extent_config:
@@ -653,7 +499,7 @@ def test_02_read_capacity16(request):
                         with target_extent_associate(target_id, extent_id):
                             iqn = f'{basename}:{target_name}'
                             with iscsi_scsi_connection(truenas_server.ip, iqn) as s:
-                                _verify_capacity(s, MB_512)
+                                verify_capacity(s, MB_512)
 
 
 def target_test_readwrite16(ip, iqn):
@@ -861,8 +707,8 @@ def test_07_report_luns(request):
                         extent_id = extent_config['id']
                         with target_extent_associate(target_id, extent_id):
                             with iscsi_scsi_connection(truenas_server.ip, iqn) as s:
-                                _verify_luns(s, [0])
-                                _verify_capacity(s, MB_100)
+                                verify_luns(s, [0])
+                                verify_capacity(s, MB_100)
                             # Now create a 512 MB zvol and associate with LUN 1
                             with zvol_dataset(zvol):
                                 with zvol_extent(zvol) as extent_config:
@@ -870,16 +716,16 @@ def test_07_report_luns(request):
                                     with target_extent_associate(target_id, extent_id, 1):
                                         # Connect to LUN 0
                                         with iscsi_scsi_connection(truenas_server.ip, iqn, 0) as s0:
-                                            _verify_luns(s0, [0, 1])
-                                            _verify_capacity(s0, MB_100)
+                                            verify_luns(s0, [0, 1])
+                                            verify_capacity(s0, MB_100)
                                         # Connect to LUN 1
                                         with iscsi_scsi_connection(truenas_server.ip, iqn, 1) as s1:
-                                            _verify_luns(s1, [0, 1])
-                                            _verify_capacity(s1, MB_512)
+                                            verify_luns(s1, [0, 1])
+                                            verify_capacity(s1, MB_512)
                             # Check again now that LUN 1 has been removed again.
                             with iscsi_scsi_connection(truenas_server.ip, iqn) as s:
-                                _verify_luns(s, [0])
-                                _verify_capacity(s, MB_100)
+                                verify_luns(s, [0])
+                                verify_capacity(s, MB_100)
 
 
 def target_test_snapshot_single_login(ip, iqn, dataset_id):
@@ -2340,21 +2186,21 @@ def test_25_resize_target_zvol(request):
             with iscsi_scsi_connection(truenas_server.ip, iqn) as s:
                 TUR(s)
                 s.blocksize = 512
-                assert MB_100 == _read_capacity16(s)
+                assert MB_100 == read_capacity16(s)
                 # Have checked using tcpdump/wireshark that a SCSI Asynchronous Event Notification
                 # gets sent 0x2A09: "CAPACITY DATA HAS CHANGED"
                 zvol_resize(zvol, MB_256)
-                assert MB_256 == _read_capacity16(s)
+                assert MB_256 == read_capacity16(s)
                 # But we can do better (in terms of test) ... turn AEN off,
                 # which means we will get a CHECK CONDITION on the next resize
                 SSH_TEST(f"echo 1 > /sys/kernel/scst_tgt/targets/iscsi/{iqn}/aen_disabled", user, password)
                 zvol_resize(zvol, MB_512)
                 expect_check_condition(s, sense_ascq_dict[0x2A09])  # "CAPACITY DATA HAS CHANGED"
-                assert MB_512 == _read_capacity16(s)
+                assert MB_512 == read_capacity16(s)
                 # Try to shrink the ZVOL again.  Expect an error
                 with pytest.raises(ValidationErrors):
                     zvol_resize(zvol, MB_256)
-                assert MB_512 == _read_capacity16(s)
+                assert MB_512 == read_capacity16(s)
 
 
 def test_26_resize_target_file(request):
@@ -2376,18 +2222,18 @@ def test_26_resize_target_file(request):
                 extent_id = config['extent']['id']
                 TUR(s)
                 s.blocksize = 512
-                assert MB_100 == _read_capacity16(s)
+                assert MB_100 == read_capacity16(s)
                 file_extent_resize(extent_id, MB_256)
-                assert MB_256 == _read_capacity16(s)
+                assert MB_256 == read_capacity16(s)
                 # Turn AEN off so that we will get a CHECK CONDITION on the next resize
                 SSH_TEST(f"echo 1 > /sys/kernel/scst_tgt/targets/iscsi/{iqn}/aen_disabled", user, password)
                 file_extent_resize(extent_id, MB_512)
                 expect_check_condition(s, sense_ascq_dict[0x2A09])  # "CAPACITY DATA HAS CHANGED"
-                assert MB_512 == _read_capacity16(s)
+                assert MB_512 == read_capacity16(s)
                 # Try to shrink the file again.  Expect an error
                 with pytest.raises(ValidationErrors):
                     file_extent_resize(extent_id, MB_256)
-                assert MB_512 == _read_capacity16(s)
+                assert MB_512 == read_capacity16(s)
 
 
 @skip_multi_initiator
@@ -2458,7 +2304,7 @@ def test_28_portal_access(request):
                 with iscsi_scsi_connection(truenas_server.ip, iqn) as s:
                     TUR(s)
                     s.blocksize = 512
-                    assert MB_100 == _read_capacity16(s)
+                    assert MB_100 == read_capacity16(s)
                 # Now, if we are in a HA config turn on ALUA and test
                 # the specific IP addresses
                 if ha:
@@ -2474,12 +2320,12 @@ def test_28_portal_access(request):
                         with iscsi_scsi_connection(truenas_server.nodea_ip, iqn) as s:
                             TUR(s)
                             s.blocksize = 512
-                            assert MB_100 == _read_capacity16(s)
+                            assert MB_100 == read_capacity16(s)
 
                         with iscsi_scsi_connection(truenas_server.nodeb_ip, iqn) as s:
                             TUR(s)
                             s.blocksize = 512
-                            assert MB_100 == _read_capacity16(s)
+                            assert MB_100 == read_capacity16(s)
 
 
 def test_29_multiple_extents():
@@ -2504,11 +2350,11 @@ def test_29_multiple_extents():
                                 with iscsi_scsi_connection(truenas_server.ip, iqn, 0) as s:
                                     TUR(s)
                                     s.blocksize = 512
-                                    assert MB_100 == _read_capacity16(s)
+                                    assert MB_100 == read_capacity16(s)
                                 with iscsi_scsi_connection(truenas_server.ip, iqn, 1) as s:
                                     TUR(s)
                                     s.blocksize = 512
-                                    assert MB_256 == _read_capacity16(s)
+                                    assert MB_256 == read_capacity16(s)
 
                                 # Now try to create another extent using the same serial number
                                 # We expect this to fail.
@@ -2697,13 +2543,13 @@ def test_32_multi_lun_targets(request):
 
     def test_target_sizes(ipaddr):
         with iscsi_scsi_connection(ipaddr, iqn1, 0) as s:
-            _verify_capacity(s, MB_100)
+            verify_capacity(s, MB_100)
         with iscsi_scsi_connection(ipaddr, iqn1, 1) as s:
-            _verify_capacity(s, MB_200)
+            verify_capacity(s, MB_200)
         with iscsi_scsi_connection(ipaddr, iqn2, 0) as s:
-            _verify_capacity(s, MB_256)
+            verify_capacity(s, MB_256)
         with iscsi_scsi_connection(ipaddr, iqn2, 1) as s:
-            _verify_capacity(s, MB_512)
+            verify_capacity(s, MB_512)
 
     with initiator_portal() as config:
         with configured_target(config, name1, 'FILE', extent_size=MB_100) as config1:
@@ -2740,7 +2586,7 @@ def test_33_no_lun_zero():
 
                                 # With libiscsi we can also check that the expected LUNs are there
                                 with iscsi_scsi_connection(truenas_server.ip, iqn, 100) as s:
-                                    _verify_luns(s, [100, 101])
+                                    verify_luns(s, [100, 101])
 
 
 def test_99_teardown(request):
